0.2.8

+++ ZYBuryPoint/ZYBuryPoint/ZYBuryPointRequestBaseModel.m
@@ -382,9 +382,9 @@ NSString *const kZYBuryPointRequestBaseModelCode = @"code";
 
 - (NSString *)devicerp
 {
-    NSInteger width = (NSInteger)(UIScreen.mainScreen.bounds.size.width * UIScreen.mainScreen.scale);
-    NSInteger height = (NSInteger)(UIScreen.mainScreen.bounds.size.height * UIScreen.mainScreen.scale);
-    return [NSString stringWithFormat:@"%ld*%ld",width,height];
+    NSInteger width = UIScreen.mainScreen.bounds.size.width * UIScreen.mainScreen.scale;
+    NSInteger height = UIScreen.mainScreen.bounds.size.height * UIScreen.mainScreen.scale;
+    return [NSString stringWithFormat:@"%ld*%ld",(long)width,(long)height];
 }
 
 - (NSString *)ipv4
